[REM-103] feat: add application manager configuration and refactor server initialization
.vscode/launch.json:
- Added configuration for Application Manager in the launch settings.

server/__init__.py:
- Refactored server initialization to streamline the application setup.
- Introduced new menu properties for better user interaction.
- Removed unused imports and code related to the previous ASGI server class.
